requirements.txt
pyiceberg[s3fs,pyarrow]>=0.7.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
boto3>=1.34.0
